load string on enter
