+++ package.json
@@ -46,7 +46,5 @@
     "webpack": "^3.0.0",
     "webpack-dev-server": "^2.5.0"
   },
-  "dependencies": {
-    "pug-html-loader": "^1.1.5"
-  }
+  "dependencies": {}
 }
